+++ cypress/ssr/ssr.spec.ts
@@ -4,6 +4,15 @@ describe('Server side rendering', () => {
   const listUrl = 'http://localhost:4200/list';
   const faviconUrl = 'http://localhost:4200/favicon.ico';
 
+  function getOrderedLifecycleHooks() {
+    return [
+      'NgxsOnInit todo',
+      'NgxsAfterBootstrap todo',
+      'NgxsOnInit lazy',
+      'NgxsAfterBootstrap lazy'
+    ];
+  }
+
   it('should make sure the Express server is running', () => {
     // Arrange & act
     cy.request(listUrl)
@@ -28,28 +37,31 @@ describe('Server side rendering', () => {
       .should('include', 'ngOnInit todo');
   });
 
+  getOrderedLifecycleHooks().forEach(hook => {
+    it(`should have '${hook}' lifecycle hook output visible`, () => {
+      // Arrange
+      // Act
+      cy.request(listUrl)
+        .its('body')
+        .should('include', hook);
+    });
+  });
+
   it('lifecycle hooks should exist in the correct order (root => lazy)', () => {
-    // Arrange & act
+    // Arrange
+    const orderedTextToFind = getOrderedLifecycleHooks();
+    // Act
     cy.request(listUrl).then(({ body }) => {
-      const ngxsOnInitIndex = body.indexOf('NgxsOnInit todo');
-      const ngxsAfterBootstrapIndex = body.indexOf('NgxsAfterBootstrap todo');
-      const ngxsOnInitLazyIndex = body.indexOf('NgxsOnInit lazy');
-      const ngxsAfterBootstrapLazyIndex = body.indexOf('NgxsAfterBootstrap lazy');
-      const stringIndexes = [
-        ngxsOnInitIndex,
-        ngxsAfterBootstrapIndex,
-        ngxsOnInitLazyIndex,
-        ngxsAfterBootstrapLazyIndex
-      ];
+      const stringIndexes = orderedTextToFind.map(text => body.indexOf(text));
 
       stringIndexes.forEach((stringIndex, index) => {
         // Assert
-        expect(stringIndex).to.be.greaterThan(-1);
-        // If it's not the first in the array
-        // every next index should more than previous
-        if (index) {
-          expect(stringIndex).to.be.greaterThan(stringIndexes[index - 1]);
-        }
+        const text = orderedTextToFind[index];
+        const previousIndex = index > 0 ? stringIndexes[index - 1] : -1;
+        expect(stringIndex).to.be.greaterThan(
+          previousIndex,
+          `'${text}' index should be greater than ${previousIndex}`
+        );
       });
     });
   });